#90 Added Markers and Lots of Features
Scatter Plot is complete
