+++ Old Scatter/scatter1.py
