+++ functions.py
@@ -9,42 +9,56 @@ toggle_switch_color='#91c153'
 # Input: id, [options]
 # Output: dcc.Dropdown
 def render_dropdown(id, options):
-    return dcc.Dropdown(id=id, options=[{'label': i, 'value': i} for i in options],
+    return dcc.Dropdown(id=id, options=[{'label': i, 'value': i} for i in options], clearable=False,
         className='card h-100' )
 
 # Function: Render drop down list
 # Input: id, [options]
 # Output: dcc.Dropdown
 def render_dropdown_dict(id, options):
-    return dcc.Dropdown(id=id, options=options,
+    return dcc.Dropdown(id=id, options=options, clearable=False,
+        className='card h-100' )
+
+# Function: Render drop down list clearable
+# Input: id, [options]
+# Output: dcc.Dropdown
+def render_dropdown_dict_clear(id, options):
+    return dcc.Dropdown(id=id, options=options, clearable=True,
         className='card h-100' )
 
 # Function: Render drop down list with selected value
 # Input: id, [options], value
 # Output: dcc.Dropdown
 def render_dropdown_valued(id, options, value):
-    return dcc.Dropdown(id=id, options=[{'label': i, 'value': i} for i in options], value=value,
+    return dcc.Dropdown(id=id, options=[{'label': i, 'value': i} for i in options], value=value, clearable=False,
+        className='card h-100' )
+
+# Function: Render drop down list with selected valueand clearable
+# Input: id, [options], value
+# Output: dcc.Dropdown
+def render_dropdown_valued_clear(id, options, value):
+    return dcc.Dropdown(id=id, options=[{'label': i, 'value': i} for i in options], value=value, clearable=True,
         className='card h-100' )
 
 # Function: Render drop down list with selected value
 # Input: id, [options], value
 # Output: dcc.Dropdown
 def render_dropdown_dict_valued(id, options, value):
-    return dcc.Dropdown(id=id, options=options, value=value,
+    return dcc.Dropdown(id=id, options=options, value=value, clearable=False,
         className='card h-100' )
 
 # Function: Render drop down list without any options
 # Input: id
 # Output: dcc.Dropdown
 def render_dropdown_blank(id):
-    return dcc.Dropdown(id=id)
+    return dcc.Dropdown(id=id, clearable=False)
 
 
 # Function: Render drop down list with label formatting (remove space between words and turn to lower case)
 # Input: id, [options]
 # Output: dcc.Dropdown
 def render_dropdown_format(id, options):
-    return dcc.Dropdown(id=id, options=[{'label': i, 'value': (i.replace(" ", "")).lower()} for i in options],
+    return dcc.Dropdown(id=id, options=[{'label': i, 'value': (i.replace(" ", "")).lower()} for i in options], clearable=False,
         className='card h-100' )
 
 # Function: Render radio items
@@ -102,13 +116,19 @@ def render_input(id, placeholder):
 # Input: id, placeholder
 # Output: dcc.Input
 def render_input_number(id, placeholder):
-    return dcc.Input(id=id, type='number', min=0, placeholder=placeholder, style={'width': '100%'})
+    return dcc.Input(id=id, type='number', min=0, placeholder=placeholder, step=0.1, style={'width': '100%'})
 
 # Function: Render number input with costum minimum value
 # Input: id, placeholder, min
 # Output: dcc.Input
 def render_input_number_min(id, placeholder, min):
     return dcc.Input(id=id, type='number', min=min, placeholder=placeholder, step=0.1, style={'width': '100%'})
+
+# Function: Render number input with costum minimum value and setting a value
+# Input: id, placeholder, min
+# Output: dcc.Input
+def render_input_number_min_value(id, min, value):
+    return dcc.Input(id=id, type='number', min=min, step=0.1, value=value, required=True, style={'width': '100%'})
 
 # Function: Render text input with delay feature, will callback after enter key pressed or input area loss its focus
 # Input: id, placeholder
@@ -155,6 +175,13 @@ def render_range_slider(id, min, max, value, step, marks):
 def render_colorpicker(id, color, r, g, b, a):
     value=dict(rgb=dict(r=r, g=g, b=b, a=a))
     return daq.ColorPicker(id=id, value=value)
+
+# Function: Render color picker
+# Input: id, min, max, value, step, label
+# Output: daq.ColorPicker
+def render_colorpicker_small(id, color, r, g, b, a):
+    value=dict(rgb=dict(r=r, g=g, b=b, a=a))
+    return daq.ColorPicker(id=id, value=value, size=150)
 
 # Function: Render numeric Input
 # Input: id, min, max, value
