--- playground/piper plot/test_pp.py
@@ -1,5 +1,0 @@
-import piper_plot as pp
-
-fig = pp.plot_background(triangle_base_length = 5)
-fig = pp.plot_point(x=[25], y=[47], fig=fig, triangle_base_length = 5)
-fig.show()
