+++ scatter01.py
@@ -10,6 +10,7 @@ import plotly.graph_objs as go
 import dash_bootstrap_components as dbc
 import colorlover as cl
 import functions as func
+import random
 
 # Initial Styles
 button_font_size='1.2em'
@@ -39,17 +40,14 @@ MARKERS_LIST = ['circle', 'square', 'diamond', 'cross', 'x', 'triangle-up', 'pen
 markers_choice = dict()
 markers_shape = dict()
 
-#external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
-
 app = dash.Dash(__name__)
-#app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
-
 
 file_name=r'..\UWA_acid_base_table.xlsx'
 
 df = pd.read_excel(file_name)
 
 cnames = df.select_dtypes(include='number').columns.values
+cat_names = df.select_dtypes(exclude=['number', 'datetime', 'datetime64']).columns.values
 
 COLORSCALES_DICT = [
     {'value': 'Blackbody', 'label': 'Blackbody'},
@@ -76,7 +74,7 @@ MARKERS_DICT = [
     {'value': 'square', 'label': 'Square'},
     {'value': 'diamond', 'label': 'Diamond'},
     {'value': 'cross', 'label': 'Cross'},
-    {'value': 'x', 'label': 'x'},
+    {'value': 'x', 'label': 'X'},
     {'value': 'triangle-up', 'label': 'Triangle-up'},
     {'value': 'pentagon', 'label': 'Pentagon'},
     {'value': 'hexagon', 'label': 'Hexagon'},
@@ -111,20 +109,12 @@ app.layout=html.Div(className='row', children=[
                             dbc.Card([
                                 dbc.CardHeader(html.H5('X Axis Value')),
                                 dbc.CardBody(children=func.render_dropdown_valued('xaxis-column', options=cnames, value=cnames[0] ))
-                            ],
-                            ),
-                            dbc.Card([
-                                dbc.CardBody(children=func.render_radio_valued('xaxis-type', options=['Linear', 'Log'], value='Linear' ))
-                            ],
+                            ], className='col-md-6', style={'margin': '0px 0px 10px 0px', 'height': '30em'}
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Y Axis Value')),
                                 dbc.CardBody(children=func.render_dropdown_valued('yaxis-column', options=cnames, value=cnames[-1] ))
-                            ]
-                            ),
-                            dbc.Card([
-                                dbc.CardBody(children=func.render_radio_valued('yaxis-type', options=['Linear', 'Log'], value='Linear' ))
-                            ], style={'height': '20em'}
+                            ], className='col-md-6', style={'margin': '0px 0px 10px 0px', 'height': '30em'}
                             ),
                         ],
                         ), id='collapse-1'
@@ -133,44 +123,49 @@ app.layout=html.Div(className='row', children=[
                 ),
                 dbc.Card([
                     dbc.CardHeader(
-                        dbc.Button('Plot Setting', id='group-2-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                        dbc.Button('Plot Settings', id='group-2-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
                     ),
                     dbc.Collapse(
                         dbc.CardBody(children=[
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Swap Axis')),
                                 dbc.CardBody(children=func.render_booleanswitch_nolab('swap', False))
-                            ],
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Grid Lines')),
                                 dbc.CardBody(children=func.render_booleanswitch_nolab('GL', True))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Zero Lines')),
                                 dbc.CardBody(children=func.render_booleanswitch_nolab('OL', False))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Label')),
                                 dbc.CardBody(children=func.render_booleanswitch_nolab('LB', False))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Legend')),
                                 dbc.CardBody(children=func.render_booleanswitch_nolab('LD', True))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Colorscale Legend')),
+                                dbc.CardHeader(html.H5('Colorbar')),
                                 dbc.CardBody(children=func.render_booleanswitch_nolab('LS', True))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Opacity')),
-                                dbc.CardBody(children=func.render_slider('opacity-slider', 0, 100, 70, 1, []))
-                            ]
+                                dbc.CardHeader(html.H5('Set X Axis Type')),
+                                dbc.CardBody(children=func.render_toggleswitch('xaxis-type', ['Linear', 'Logarithmic'], False))
+                            ], className='col-md-6',
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Set Y Axis Type')),
+                                dbc.CardBody(children=func.render_toggleswitch('yaxis-type', ['Linear', 'Logarithmic'], False))
+                            ], className='col-md-6',
                             ),
                         ],
                         ), id='collapse-2'
@@ -179,39 +174,39 @@ app.layout=html.Div(className='row', children=[
                 ),
                 dbc.Card([
                     dbc.CardHeader(
-                        dbc.Button('Customised Lines', id='group-3-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                        dbc.Button('Thresholds', id='group-3-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
                     ),
                     dbc.Collapse(
                         dbc.CardBody(children=[
                             dbc.Card([
                                 dbc.CardHeader(html.H5('X Axis Threshold')),
                                 dbc.CardBody(children=func.render_input_number('X-thredshold', 'X Axis Threshold'))
-                            ],
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Y Axis Threshold')),
                                 dbc.CardBody(children=func.render_input_number('Y-thredshold', 'Y Axis Threshold'))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Linear Best Fit Line')),
-                                dbc.CardBody(children=func.render_booleanswitch_nolab('linear', False))
-                            ]
+                                dbc.CardHeader(html.H5('X Threshold Line Style')),
+                                dbc.CardBody(children=func.render_dropdown_dict_valued('x-threshold-style', DASH_DICT, 'solid'))
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Best Fit Line Style')),
-                                dbc.CardBody(children=func.render_dropdown_dict('change-dash', DASH_DICT))
-                            ]
+                                dbc.CardHeader(html.H5('Y Threshold Line Style')),
+                                dbc.CardBody(children=func.render_dropdown_dict_valued('y-threshold-style', DASH_DICT, 'solid'))
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('X Tick Distance')),
-                                dbc.CardBody(children=func.render_input_number_min('X-dtick', 'X Axis Delta Tick', 0))
-                            ]
+                                dbc.CardHeader(html.H5('X Threshold Line Color')),
+                                dbc.CardBody(children=func.render_colorpicker_small('x-threshold-color', '#ffffff', 0, 0, 0, 1))
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Y Tick Distance')),
-                                dbc.CardBody(children=func.render_input_number_min('Y-dtick', 'Y Axis Delta Tick', 0))
-                            ]
+                                dbc.CardHeader(html.H5('Y Threshold Line Color')),
+                                dbc.CardBody(children=func.render_colorpicker_small('y-threshold-color', '#ffffff', 0, 0, 0, 1))
+                            ], className='col-md-6',
                             ),
                         ],
                         ), id='collapse-3'
@@ -220,37 +215,149 @@ app.layout=html.Div(className='row', children=[
                 ),
                 dbc.Card([
                     dbc.CardHeader(
-                        dbc.Button('Plot Style', id='group-4-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                        dbc.Button('Statistics', id='group-4-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
                     ),
                     dbc.Collapse(
                         dbc.CardBody(children=[
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Color and Hover Text Group')),
-                                dbc.CardBody(children=func.render_dropdown_valued('color-drop', df.columns, df.columns[0]))
-                            ],
+                                dbc.CardHeader(html.H5('Linear Best Fit Line')),
+                                dbc.CardBody(children=func.render_booleanswitch_nolab('linear', False))
+                            ]
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Select Sub-group')),
-                                dbc.CardBody(children=func.render_dropdown_blank('selected-groupby'))
-                            ],
+                                dbc.CardHeader(html.H5('Best Fit Line Style')),
+                                dbc.CardBody(children=func.render_dropdown_dict('change-dash', DASH_DICT))
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Sub-group Marker Style')),
+                                dbc.CardHeader(html.H5('Best Fit Line Color')),
+                                dbc.CardBody(children=func.render_colorpicker_small('fit-color-picker', '#ffffff', 0, 0, 0, 1))
+                            ], className='col-md-6',
+                            ),
+                        ],
+                        ), id='collapse-4'
+                    ),
+                ], color=cardbody_color, outline=True, style={'font-size': cardbody_font_size}
+                ),
+                dbc.Card([
+                    dbc.CardHeader(
+                        dbc.Button('Marker Style', id='group-5-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                    ),
+                    dbc.Collapse(
+                        dbc.CardBody(children=[
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Category Marker Style')),
                                 dbc.CardBody(children=func.render_dropdown_dict_valued('alignment-markers-dropdown', MARKERS_DICT, 'circle'))
-                            ],
+                            ]
+                            ),
+                            dbc.Card([
+                                dbc.CardBody(children=func.render_booleanswitch('marker-style-tog', 'Set Marker by a Group', False))
+                            ]
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Group By')),
+                                dbc.CardBody(children=func.render_dropdown_valued('marker-drop', cat_names, cat_names[0]))
+                            ], className='col-md-6', style={'margin': '0px 0px 10px 0px', 'height': '30em'}
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Select Category')),
+                                dbc.CardBody(children=func.render_dropdown_blank('marker-selected-groupby'))
+                            ], className='col-md-6', style={'margin': '0px 0px 10px 0px', 'height': '30em'}
+                            ),
+                        ],
+                        ), id='collapse-5'
+                    ),
+                ], color=cardbody_color, outline=True, style={'font-size': cardbody_font_size}
+                ),
+                dbc.Card([
+                    dbc.CardHeader(
+                        dbc.Button('Marker Size', id='group-6-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                    ),
+                    dbc.Collapse(
+                        dbc.CardBody(children=[
+                            dbc.Card([
+                                dbc.CardBody(children=func.render_booleanswitch('marker-size-tog', 'Set Marker Size by a Group', False))
+                            ]
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('The Maximum Size for the Markers')),
+                                dbc.CardBody(children=func.render_input_number_min_value('marker-max', 1, 15))
+                            ]
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Marker Size Group')),
+                                dbc.CardBody(children=func.render_dropdown_valued('marker-size', cnames, cnames[0]))
+                            ], style={'margin': '0px 0px 10px 0px', 'height': '30em'}
+                            ),
+                        ],
+                        ), id='collapse-6'
+                    ),
+                ], color=cardbody_color, outline=True, style={'font-size': cardbody_font_size}
+                ),
+                dbc.Card([
+                    dbc.CardHeader(
+                        dbc.Button('Plot Color', id='group-7-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                    ),
+                    dbc.Collapse(
+                        dbc.CardBody(children=[
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Group By')),
+                                dbc.CardBody(children=func.render_dropdown_valued('color-drop', df.columns, df.columns[0]))
+                            ]
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Select Category')),
+                                dbc.CardBody(children=func.render_dropdown_blank('color-selected-groupby'))
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
                                 dbc.CardHeader(html.H5('Change Colorscale')),
                                 dbc.CardBody(children=func.render_dropdown_dict_valued('alignment-colorscale-dropdown', COLORSCALES_DICT, 'Greys'))
-                            ]
+                            ], className='col-md-6',
                             ),
                             dbc.Card([
-                                dbc.CardHeader(html.H5('Sub-group Color')),
-                                dbc.CardBody(children=func.render_colorpicker('my-color-picker', '#ffffff', 22, 222, 160, 1))
-                            ],
+                                dbc.CardHeader(html.H5('Category Color')),
+                                dbc.CardBody(children=func.render_colorpicker_small('my-color-picker', '#ffffff', 22, 222, 160, 1))
+                            ]
                             ),
                         ],
-                        ), id='collapse-4'
+                        ), id='collapse-7'
+                    ),
+                ], color=cardbody_color, outline=True, style={'font-size': cardbody_font_size}
+                ),
+                dbc.Card([
+                    dbc.CardHeader(
+                        dbc.Button('Graph Settings', id='group-8-toggle', color=cardheader_color, style={'font-size': button_font_size}, block=True, )
+                    ),
+                    dbc.Collapse(
+                        dbc.CardBody(children=[
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Graph Height')),
+                                dbc.CardBody(children=func.render_slider('graph-height', 600, 1200, 600, 50, [600, 700, 800, 900, 1000, 1100, 1200]), style={'padding':'5% 5% 10% 5%'})
+                            ], style={'width': '100%'}
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Graph Width')),
+                                dbc.CardBody(children=func.render_slider('graph-width', 800, 1400, 800, 50, [800, 900, 1000, 1100, 1200, 1300, 1400]), style={'padding':'5% 5% 10% 5%'})
+                            ], style={'width': '100%'}
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Opacity')),
+                                dbc.CardBody(children=func.render_slider('opacity-slider', 0, 100, 70, 1, []))
+                            ], style={'width': '100%'}
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('X Tick Distance')),
+                                dbc.CardBody(children=func.render_input_number_min('X-dtick', 'X Axis Delta Tick', 0))
+                            ], className='col-md-6',
+                            ),
+                            dbc.Card([
+                                dbc.CardHeader(html.H5('Y Tick Distance')),
+                                dbc.CardBody(children=func.render_input_number_min('Y-dtick', 'Y Axis Delta Tick', 0))
+                            ], className='col-md-6',
+                            ),
+                        ],
+                        ), id='collapse-8'
                     ),
                 ], color=cardbody_color, outline=True, style={'font-size': cardbody_font_size}
                 ),
@@ -273,11 +380,11 @@ app.layout=html.Div(className='row', children=[
 
 # Accordion Toggle Callback
 @app.callback(
-    [Output(f'collapse-{i}', 'is_open') for i in range(1,5)],
-    [Input(f'group-{i}-toggle', 'n_clicks') for i in range(1,5)],
-    [State(f'collapse-{i}', 'is_open') for i in range(1,5)]
+    [Output(f'collapse-{i}', 'is_open') for i in range(1,9)],
+    [Input(f'group-{i}-toggle', 'n_clicks') for i in range(1,9)],
+    [State(f'collapse-{i}', 'is_open') for i in range(1,9)]
 )
-def toggle_accordion(n1, n2, n3, n4, is_open1, is_open2, is_open3, is_open4):
+def toggle_accordion(n1, n2, n3, n4, n5, n6, n7, n8, is_open1, is_open2, is_open3, is_open4, is_open5, is_open6, is_open7, is_open8):
     ctx = dash.callback_context
 
     if not ctx.triggered:
@@ -286,14 +393,22 @@ def toggle_accordion(n1, n2, n3, n4, is_open1, is_open2, is_open3, is_open4):
         button_id = ctx.triggered[0]['prop_id'].split('.')[0]
 
     if button_id ==  'group-1-toggle' and n1:
-        return not is_open1, False, False, False
+        return not is_open1, False, False, False, False, False, False, False
     elif button_id ==  'group-2-toggle' and n2:
-        return False, not is_open2, False, False
+        return False, not is_open2, False, False, False, False, False, False
     elif button_id ==  'group-3-toggle' and n3:
-        return False, False, not is_open3, False
+        return False, False, not is_open3, False, False, False, False, False
     elif button_id ==  'group-4-toggle' and n4:
-        return False, False, False, not is_open4
-    return False, False, False, False
+        return False, False, False, not is_open4, False, False, False, False
+    elif button_id ==  'group-5-toggle' and n5:
+        return False, False, False, False, not is_open5, False, False, False
+    elif button_id ==  'group-6-toggle' and n6:
+        return False, False, False, False, False, not is_open6, False, False
+    elif button_id ==  'group-7-toggle' and n7:
+        return False, False, False, False, False, False, not is_open7, False
+    elif button_id ==  'group-8-toggle' and n8:
+        return False, False, False, False, False, False, False, not is_open8
+    return False, False, False, False, False, False, False, False
 
 @app.callback(
     Output('my-color-picker', 'value'),
@@ -323,11 +438,34 @@ def show_fitline(xv, yv):
 		return False,False
 
 @app.callback(
-	[Output('selected-groupby', 'options'),
+    [Output('x-threshold-style', 'disabled'),
+    Output('y-threshold-style', 'disabled'),
+    Output('x-threshold-color', 'disabled'),
+    Output('y-threshold-color', 'disabled')],
+    [Input('X-thredshold', 'value'),
+     Input('Y-thredshold', 'value'),]
+)
+def enable_threshold_styles(x_t, y_t):
+    if x_t is not None:
+        x_style = False
+        x_color= False
+    else:
+        x_style = True
+        x_color = True
+    if y_t is not None:
+        y_style = False
+        y_color = False
+    else:
+        y_style = True
+        y_color = True
+    return x_style, y_style, x_color, y_color
+
+@app.callback(
+	[Output('color-selected-groupby', 'options'),
 	Output('alignment-colorscale-dropdown', 'disabled'),
 	Output('my-color-picker','disabled'),
 	Output('LS', 'disabled'),
-	Output('selected-groupby', 'disabled')
+	Output('color-selected-groupby', 'disabled')
 	],
 	[Input('color-drop', 'value'),]
 	)
@@ -336,21 +474,39 @@ def traces_groupby(color_drop):
 		idx =0
 		for i in df[color_drop].unique():
 			markers_choice[i] = default_color[idx % num_of_color]
-			markers_shape[i] = random.choice(MARKERS_LIST)
 			idx += 1
 		return [{'label': i, 'value': i} for i in df[color_drop].unique()], True, False, True, False
 	else:
 		return [{'label': color_drop, 'value': color_drop}], False, True, False, True
 
 @app.callback(
-	Output('change-dash', 'disabled'),
+    [Output('marker-selected-groupby', 'options'),
+    Output('marker-selected-groupby', 'disabled'),
+    Output('marker-drop', 'disabled')],
+    [Input('marker-style-tog', 'on'),
+    Input('marker-drop', 'value')]
+)
+def markers_groupby(marker_on, marker_drop):
+    idx = 0
+    if marker_on:
+        for i in df[marker_drop].unique():
+            markers_shape[i] = random.choice(MARKERS_LIST)
+            idx += 1
+        return [{'label': i, 'value': i} for i in df[marker_drop].unique()], False, False
+    else:
+        return [{'label': marker_drop, 'value': marker_drop}], True, True
+
+
+@app.callback(
+	[Output('change-dash', 'disabled'),
+    Output('fit-color-picker', 'disabled')],
 	[Input('linear', 'on'),]
 	)
 def show_linear(on):
 	if on:
-		return False
+		return False, False
 	else:
-		return True
+		return True, True
 
 @app.callback(
     Output('indicator-graphic', 'figure'),
@@ -371,16 +527,35 @@ def show_linear(on):
      Input('Y-dtick', 'value'),
      Input('X-thredshold', 'value'),
      Input('Y-thredshold', 'value'),
-     Input('selected-groupby', 'value'),
+     Input('color-selected-groupby', 'value'),
      Input('my-color-picker', 'value'),
      Input('LB', 'on'),
      Input('LS', 'on'),
-     Input('change-dash', 'value')])
+     Input('change-dash', 'value'),
+     Input('graph-height', 'value'),
+     Input('graph-width', 'value'),
+     Input('fit-color-picker', 'value'),
+     Input('x-threshold-style', 'value'),
+     Input('y-threshold-style', 'value'),
+     Input('x-threshold-color', 'value'),
+     Input('y-threshold-color', 'value'),
+     Input('marker-size-tog', 'on'),
+     Input('marker-size', 'value'),
+     Input('marker-max', 'value'),
+     Input('marker-drop', 'value'),
+     Input('marker-selected-groupby', 'value'),
+     Input('marker-style-tog', 'on'),
+     ])
 def update_graph(xaxis_column_name, yaxis_column_name,
                  xaxis_type, yaxis_type,
                  alignment_colorscale_dropdown, 
                  swap, linear, GL, OL, 
-                 alignment_markers_dropdown, color_var, LD, OS, X_D, Y_D, X_T, Y_T, G_t, C_P, LB, LS, CD):
+                 alignment_markers_dropdown, color_var, 
+                 LD, OS, X_D, Y_D, X_T, Y_T, G_t, C_P, LB,
+                 LS, CD, graph_height, graph_width, 
+                 fit_color_picker, x_t_style, y_t_style,
+                 x_t_color, y_t_color, use_size, marker_size, marker_max, 
+                 marker_drop, marker_groupby, marker_on):
 
     if swap:
     	# Swapping the x and y axes names and values
@@ -393,7 +568,17 @@ def update_graph(xaxis_column_name, yaxis_column_name,
 
 
     threshold_shape = []
+    x_color = 'rgba({}, {}, {}, {})'.format(
+        x_t_color['rgb']['r'],
+        x_t_color['rgb']['g'],
+        x_t_color['rgb']['b'],
+        x_t_color['rgb']['a'],)
 
+    y_color = 'rgba({}, {}, {}, {})'.format(
+        y_t_color['rgb']['r'],
+        y_t_color['rgb']['g'],
+        y_t_color['rgb']['b'],
+        y_t_color['rgb']['a'],)
 
     #if users set a threshold for X, then show this line
     if X_T !=None:
@@ -402,7 +587,11 @@ def update_graph(xaxis_column_name, yaxis_column_name,
     	x0=X_T,
     	x1=X_T,
     	y0=df[yaxis_column_name].min(),
-    	y1=df[yaxis_column_name].max()
+    	y1=df[yaxis_column_name].max(),
+        line=dict(
+            color=x_color,
+            dash=x_t_style
+        )
     	))
 
     #if users set a threshold for Y, then show this line
@@ -412,7 +601,11 @@ def update_graph(xaxis_column_name, yaxis_column_name,
     	x0=df[xaxis_column_name].min(),
     	x1=df[xaxis_column_name].max(),
     	y0=Y_T,
-    	y1=Y_T
+    	y1=Y_T,
+        line=dict(
+            color=y_color,
+            dash=y_t_style
+        )
     	))}
 
     mode_t='markers'
@@ -430,15 +623,25 @@ def update_graph(xaxis_column_name, yaxis_column_name,
     		if color_var is not None:
     			if i == G_t:
     				markers_choice[i] = picker_markers_color
-    				markers_shape[i] = alignment_markers_dropdown
+
+    if marker_on:
+        for cat in df[marker_drop].unique():
+            if marker_drop is not None:
+                if cat == marker_groupby:
+                    markers_shape[cat] = alignment_markers_dropdown
+    
+    if use_size:
+        marker_data = np.array(df[marker_size], dtype='float64')
+        marker_size = list(np.interp(marker_data, (marker_data.min(), marker_data.max()), (1, marker_max)))
+    else:
+        marker_size=15
 
     traces = []
     #if the data type of the group by column is object, it will show each different values with different colors
-    if df[color_var].dtypes=='object':
+    if df[color_var].dtypes=='object' and not marker_on:
     	for i in df[color_var].unique():
     		df_by = df[df[color_var] == i]
-    		if i == G_t:
-    			traces.append(go.Scatter(
+    		traces.append(go.Scatter(
 	    		x=df_by[xaxis_column_name],
 	       		y=df_by[yaxis_column_name],
 	       		mode=mode_t,	       		
@@ -446,33 +649,36 @@ def update_graph(xaxis_column_name, yaxis_column_name,
 	       		textposition='top center',
 	       		opacity=OS/100,
 	       		marker={
-	       			'size': 15,
+	       			'size': marker_size,
 	       			'line': {'width' :0.5, 'color': 'white'},
 	       			'symbol': alignment_markers_dropdown,
 	       			'color': markers_choice[i],
-	       			'symbol':markers_shape[i]
 	       		},
 	       		name=i
-	       		))
-    		else:
-		    	traces.append(go.Scatter(
-		    		x=df_by[xaxis_column_name],
-		       		y=df_by[yaxis_column_name],
-		       		mode=mode_t,
-		       		text=i,
-		       		textposition='top center',
-		       		opacity=OS/100,
-		       		marker={
-		       			'size': 15,
-		       			'line': {'width' :0.5, 'color': 'white'},
-		       			'color': markers_choice[i],
-		       			'symbol':markers_shape[i]
-		       		},
-		       		name=i
-		    	))
-
+	       	    )
+            )
+    elif df[color_var].dtypes=='object' and marker_on:
+        for i in df[color_var].unique():
+            for j in df[marker_drop].unique():
+                df_by = df[df[color_var] == i][df[marker_drop] == j]
+                traces.append(go.Scatter(
+                    x=df_by[xaxis_column_name],
+                    y=df_by[yaxis_column_name],
+                    mode=mode_t,
+                    text=i,
+                    textposition='top center',
+                    opacity=OS/100,
+                    marker={
+                        'size': marker_size,
+                        'line': {'width' :0.5, 'color': 'white'},
+                        'color': markers_choice[i],
+                        'symbol':markers_shape[j]
+                        },
+                        name=str(i) + ' : ' + str(j)
+                    )
+                )
     #if the data type of the group by column is int or float, it will show the VS, and the color based on the X values
-    if df[color_var].dtypes=='int64' or df[color_var].dtypes=='float64':
+    elif (df[color_var].dtypes=='int64' or df[color_var].dtypes=='float64') and not marker_on:
 	    traces.append(go.Scatter(
 	    	x=df[xaxis_column_name],
 	    	y=df[yaxis_column_name],
@@ -480,7 +686,7 @@ def update_graph(xaxis_column_name, yaxis_column_name,
 	    	text = color_var,
 	    	opacity=OS/100,
 	    	marker=dict(
-	    		size = 15,
+	    		size = marker_size,
 	        	line = {'width': 0.5, 'color': 'white'},
 	        	color = df[color_var],
 	        	colorscale = alignment_colorscale_dropdown,
@@ -492,7 +698,37 @@ def update_graph(xaxis_column_name, yaxis_column_name,
 	    	),
 	    	name='{} VS {}'.format(xaxis_column_name, yaxis_column_name)
 	    ))
-
+    elif (df[color_var].dtypes=='int64' or df[color_var].dtypes=='float64') and marker_on:
+        for cat in df[marker_drop].unique():
+            df_by = df[df[marker_drop] == cat]
+            showScale = False
+            if cat == df[marker_drop].unique()[-1]:
+                showScale = LS
+            traces.append(go.Scatter(
+                x=df_by[xaxis_column_name],
+                y=df_by[yaxis_column_name],
+                mode=mode_t,
+                text = color_var,
+                opacity=OS/100,
+                marker=dict(
+                    size = marker_size,
+                    line = {'width': 0.5, 'color': 'white'},
+                    color = df[color_var],
+                    colorscale = alignment_colorscale_dropdown,
+                    colorbar=dict(
+                        title=color_var
+                        ),
+                    showscale = showScale,
+                    symbol = markers_shape[cat]
+                ),
+                name=cat,
+            ))
+        
+    fit_color = 'rgba({}, {}, {}, {})'.format(
+        fit_color_picker['rgb']['r'],
+        fit_color_picker['rgb']['g'],
+        fit_color_picker['rgb']['b'],
+        fit_color_picker['rgb']['a'],)
     traces.append(go.Scatter(
     	x=df[xaxis_column_name],
         y=line,
@@ -502,7 +738,7 @@ def update_graph(xaxis_column_name, yaxis_column_name,
             size = 15,
             opacity = 0.5,
             line = {'width': 0.5, 'color': 'white'},
-            color = 'black',
+            color = fit_color,
             showscale = LD
         ),
         line = dict(
@@ -514,14 +750,14 @@ def update_graph(xaxis_column_name, yaxis_column_name,
     layou_t=dict(
     	xaxis={
             'title': xaxis_column_name,
-            'type': 'linear' if xaxis_type == 'Linear' else 'log',
+            'type': 'log' if xaxis_type else 'linear',
            	'showgrid': GL,
            	'zeroline': OL,
             'dtick': X_D
         },
         yaxis={
             'title': yaxis_column_name,
-            'type': 'linear' if yaxis_type == 'Linear' else 'log',
+            'type': 'log' if yaxis_type else 'linear',
             'showgrid': GL,
             'zeroline': OL,
             'dtick': Y_D
@@ -529,7 +765,9 @@ def update_graph(xaxis_column_name, yaxis_column_name,
         title= xaxis_column_name + ' vs. ' + yaxis_column_name,
         showlegend = LD,
         shapes=threshold_shape,
-        hovermode='closest'
+        hovermode='closest',
+        height=graph_height,
+        width=graph_width
     	)
 
     return {
